+++ Exercicios06-03-26/Exercicio18.py
@@ -1,4 +1,4 @@
-preço= int(input('Digite o preço do produto: '))
+preço= float(input('Digite o preço do produto: '))
 
 desconto= preço * 0.10
 valfinal= preço - desconto
