+++ Exercicios06-03-26/Exercicio8.py
@@ -1,4 +1,4 @@
-x=int(input('Digite um numero: '))
+x=float(input('Digite um numero: '))
 
 quadrado= x * x
